[Feature] Instances!!
Write a component in another gnd file and make it appear in the main one!
Specified by
```
>fileName(x, y)
/
```
This change also moved the actionables definition to the parser which should theoretically speed up the drawing speed in turn increasing the maximum complexity.
An Instance has access to the main's variables and uses the same functions as the main file!
On the user side there has been a modification to Main and MConfig requiring the path to be passed to GND.